Automate OCP creation for OCS CiCD
A new framework is provided that parameterizes most of the
user specified fields in the control file var.tfvars.  This
framework is composed of a few scripts and environment variables.

A create script is provided that allows the user to specify the
OCP version, the number of worker nodes, and the amount of CPU
and memory to be allocated for each worker node.  The user can
similarly control the amount of memory and CPU to be allocated
for master nodes.  These environment variables are optional.
The default behavior is so to use today's hadcorded parameters,
so the new scripts are transparent to the end user.

The only difference in default behavior is that libvirt images
are allocated in /var/lib/libvirt instead of /home/libvirt/images,
but this difference can be eliminated as it is subject to an
environment varible.  See the scripts/README.md for details.

A setup KVM host script is provided to install and configure
libvirt, firewalls, dns overlays, iptables, and HAProxy enabling
the user to remotely connect to the OpenShift webconsole.

A final script is provided to destroy the cluster from the
host server.  This scripts removes all virsh objects associated
with the cluster.

The user can create a cluster at OCP 4.4, tear it down, and
create another cluster at OCP 4.5, ... OCP 4.6 is also supported.

This support does not impact the existing use of the project.
It simply defines an alternate use that can be scripted.

Signed-off-by: Luke Browning <lukebrowning@us.ibm.com>